Consolidate benchmarking for Go hash
